All versionUpgrade-related classes moved to versionupgrade package #302
