+++ src/main/groovy/org/shipkit/internal/gradle/versionupgrade/VersionUpgradeConsumerExtension.java
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle;
+package org.shipkit.internal.gradle.versionupgrade;
 
 import java.io.File;
