+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.version-upgrade-consumer.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.VersionUpgradeConsumerPlugin
+implementation-class=org.shipkit.internal.gradle.versionupgrade.VersionUpgradeConsumerPlugin
