+++ src/test/groovy/org/shipkit/internal/gradle/versionupgrade/DependencyNewVersionParserTest.groovy
@@ -1,6 +1,5 @@
 package org.shipkit.internal.gradle.versionupgrade
 
-import org.shipkit.internal.gradle.VersionUpgradeConsumerExtension
 import spock.lang.Specification
 
 class DependencyNewVersionParserTest extends Specification {
