+++ src/test/groovy/org/shipkit/internal/gradle/versionupgrade/VersionUpgradeConsumerPluginTest.groovy
@@ -1,11 +1,9 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.gradle.versionupgrade
 
 import org.gradle.api.tasks.Exec
 import org.shipkit.gradle.git.GitPushTask
 import org.shipkit.internal.gradle.configuration.DeferredConfiguration
 import org.shipkit.internal.gradle.git.GitCheckOutTask
-import org.shipkit.internal.gradle.versionupgrade.CreatePullRequestTask
-import org.shipkit.internal.gradle.versionupgrade.ReplaceVersionTask
 import testutil.PluginSpecification
 
 class VersionUpgradeConsumerPluginTest extends PluginSpecification {
